requirements.txt
Django==1.6.5
django-haystack==2.1.0
Whoosh==2.6.0